userClass
Each time a new account is created, a user class should be created in association with the new account. Meaning that each user can keep track of their push up pull up and sit up scores as well as a total. This makes it easier get these values when needed.